interaction code split
